skills/sf-ai-agentforce-observability/validation/requirements.txt
# TDD Validation Framework Dependencies
# Install: pip install -r validation/requirements.txt

# Testing framework
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0

# CLI testing (Click's built-in test runner)
click>=8.1.0

# Data handling (same as main skill)
polars>=1.0.0
pyarrow>=15.0.0

# HTTP mocking for offline tests
responses>=0.25.0
pytest-httpx>=0.30.0

# JSON schema validation
jsonschema>=4.21.0

# Report generation
rich>=13.0.0
tabulate>=0.9.0

# Type checking (optional)
mypy>=1.8.0
